backend/requirements.txt
# ================================================
# CRACLE - Backend Dependencies
# Cognitive Reasoning and Adaptive Learning Engine
# ================================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Azure SDKs
azure-identity==1.19.0
azure-ai-projects==1.0.0b7
azure-ai-inference==1.0.0b7
azure-cosmos==4.9.0
azure-storage-blob==12.24.0
azure-monitor-opentelemetry==1.6.4
openai==1.59.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data & ML
pydantic==2.10.4
pydantic-settings==2.7.0
numpy==2.2.1
pandas==2.2.3

# Monitoring / Telemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-client==0.21.1

# MCP (Model Context Protocol)
mcp==1.2.0

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
Jinja2==3.1.5
